Refactored to make better sense
